Autorization

+++ App/controllers/ListingController.php
@@ -4,6 +4,8 @@ namespace App\Controllers;
 
 use Framework\Database;
 use Framework\Validation;
+use Framework\Session;
+use Framework\Authorization;
 
 class ListingController
 {
@@ -17,7 +19,7 @@ class ListingController
 
   public function index()
   {
-    $listings = $this->db->query('SELECT * FROM `listings`')->fetchAll();
+    $listings = $this->db->query('SELECT * FROM `listings` ORDER BY created_at DESC')->fetchAll();
     //inspect($listings);
 
     loadView('listings/index', [
@@ -61,7 +63,7 @@ class ListingController
     $allowedFields = ['company', 'city', 'state', 'address', 'phone', 'email', 'title', 'role_summary', 'description', 'requirements', 'benefits', 'tags', 'salary', 'work_location'];
 
     $newListingData = array_intersect_key($_POST, array_flip($allowedFields));
-    $newListingData['user_id'] = 1;
+    $newListingData['user_id'] = Session::get('user')['id'];
 
     //sanitize data
     $newListingData = array_map('sanitize', $newListingData);
@@ -147,6 +149,17 @@ class ListingController
       return;
     }
 
+    //Authorization
+    // if (Session::get('user')['id'] !== $listing->user_id) {
+    //   $_SESSION['error_message'] = 'You are not authorized to delete this listing';
+    //   return redirect('/listings/'. $listing->id);
+    // }
+
+    if (!Authorization::isOwner($listing->user_id)) {
+      $_SESSION['error_message'] = 'You are not authorized to delete this listing';
+      return redirect('/listings/' . $listing->id);
+    }
+
     $this->db->query('DELETE FROM `listings` WHERE `id` = :id', $params);
     // Set flash msg
     $_SESSION['success_message'] = 'Listing deleted successfully';
@@ -173,6 +186,12 @@ class ListingController
       ErrorController::notFound('Listing not found');
       return;
     }
+
+    if (!Authorization::isOwner($listing->user_id)) {
+      $_SESSION['error_message'] = 'You are not authorized to update this listing';
+      return redirect('/listings/' . $listing->id);
+    }
+
     loadView('listings/edit', [
       'listing' => $listing
     ]);
@@ -196,6 +215,11 @@ class ListingController
     if (!$listing) {
       ErrorController::notFound('Listing not found');
       return;
+    }
+
+    if (!Authorization::isOwner($listing->user_id)) {
+      $_SESSION['error_message'] = 'You are not authorized to update this listing';
+      return redirect('/listings/' . $listing->id);
     }
 
     $allowedFields = ['company', 'city', 'state', 'address', 'phone', 'email', 'title', 'role_summary', 'description', 'requirements', 'benefits', 'tags', 'salary', 'work_location'];
@@ -240,11 +264,11 @@ class ListingController
         $updateFields[] = "{$field} = :{$field}";
       }
 
-      $updateFields = implode(', ',$updateFields);
+      $updateFields = implode(', ', $updateFields);
       //inspectAndDie($updateFields);
       $updateQuery = "UPDATE `listings` SET $updateFields WHERE `id`=:id";
       //inspectAndDie($updateQuery);
-      $updatedValues['id'] =$id;
+      $updatedValues['id'] = $id;
       $this->db->query($updateQuery, $updatedValues);
 
       $_SESSION['success_message'] = 'Listing Updated';
